+++ changelog.md
@@ -4,6 +4,7 @@
 
 ### New Features
 
+- #909 Allow to control the product availability based on the existence of an attribute combination.
 - Affiliates
   - #896 Added a cart rule for affiliates.
   - Added a button to remove the assignment of a customer to an affiliate on customer edit page.